feat: exstract mock data
