+++ Cargo.toml
@@ -14,6 +14,7 @@ ark-bn254 = { version = "0.4.0", features = ["curve"] }
 tiny-keccak = { version = "2.0", features = ["sha3","keccak"] }
 num-bigint = "0.4.4"
 num-traits = "0.2.18"
+once_cell = "1.19.0"
 
 [[example]]
 name = "fflonk_verifier_examples"
