+++ src/lib.rs
@@ -3,10 +3,11 @@ pub mod compute_fej;
 pub mod compute_pi;
 pub mod compute_r;
 pub mod inversion;
+mod mock;
 pub mod pairing;
 pub mod proof;
 pub mod test;
 pub mod verifier;
 pub mod vk;
 
-pub use crate::proof::{get_domain_size, get_omegas, get_pubSignals, padd_bytes32, Omegas, Proof};
+pub use crate::proof::{get_omegas, padd_bytes32, Omegas, Proof};
