+++ src/mock.rs
@@ -1,0 +1,37 @@
+use ark_bn254::Fr;
+use once_cell::unsync::Lazy;
+use std::str::FromStr;
+
+pub static MOCK_PUB_INPUT: &str =
+    "14516932981781041565586298118536599721399535462624815668597272732223874827152";
+pub static MOCK_PROOF_DATA: [&str; 24] = [
+    "12195165594784431822497303968938621279445690754376121387655513728730220550454",
+    "19482351300768228183728567743975524187837254971200066453308487514712354412818",
+    "270049702185508019342640204324826241417613526941291105097079886683911146886",
+    "8044577183782099118358991257374623532841698893838076750142877485824795072127",
+    "18899554350581376849619715242908819289791150067233598694602356239698407061017",
+    "868483199604273061042760252576862685842931472081080113229115026384087738503",
+    "15400234196629481957150851143665757067987965100904384175896686561307554593394",
+    "1972554287366869807517068788787992038621302618305780153544292964897315682091",
+    "13012702442141574024514112866712813523553321876510290446303561347565844930654",
+    "6363613431504422665441435540021253583148414748729550612486380209002057984394",
+    "16057866832337652851142304414708366836077577338023656646690877057031251541947",
+    "12177497208173170035464583425607209406245985123797536695060336171641250404407",
+    "1606928575748882874942488864331180511279674792603033713048693169239812670017",
+    "12502690277925689095499239281542937835831064619179570213662273016815222024218",
+    "21714950310348017755786780913378098925832975432250486683702036755613488957178",
+    "7373645520955771058170141217317033724805640797155623483741097103589211150628",
+    "10624974841759884514517518996672059640247361745924203600968035963539096078745",
+    "12590031312322329503809710776715067780944838760473156014126576247831324341903",
+    "17676078410435205056317710999346173532618821076911845052950090109177062725036",
+    "13810130824095164415807955516712763121131180676617650812233616232528698737619",
+    "9567903658565551430748252507556148460902008866092926659415720362326593620836",
+    "17398514793767712415669438995039049448391479578008786242788501594157890722459",
+    "11804645688707233673914574834599506530652461017683048951953032091830492459803",
+    "6378827379501409574366452872421073840754012879130221505294134572417254316105",
+];
+
+// TODO: optimize
+pub fn get_domain_size() -> Fr {
+    Fr::from_str("16777216").unwrap()
+}
